Refactor type handling and improve compiler compatibility
- Updated various source files to replace `int` with `size_t` for loop indices and size calculations, enhancing type safety and preventing potential overflow issues.
- Introduced `__attribute__((fallthrough))` for better fallthrough handling in switch cases, ensuring compliance with modern compiler standards.
- Adjusted conditional compilation directives to maintain compatibility across different compilers and configurations.
- Enhanced error handling in several functions to ensure robust behavior when dealing with size and index checks.

These changes collectively improve code quality, maintainability, and compatibility with modern C standards.

+++ src/client/cl_parse.c
@@ -260,7 +260,7 @@ static void CL_ParseSnapshot( msg_t *msg ) {
 	// read areamask
 	newSnap.areabytes = MSG_ReadByte( msg );
 
-	if ( newSnap.areabytes > sizeof(newSnap.areamask) )
+	if ( (size_t)newSnap.areabytes > sizeof(newSnap.areamask) )
 	{
 		Com_Error( ERR_DROP,"CL_ParseSnapshot: Invalid size %d for areamask", newSnap.areabytes );
 		return;
@@ -410,7 +410,7 @@ void CL_SystemInfoChanged( qboolean onlyGame ) {
 	// scan through all the variables in the systeminfo and locally set cvars to match
 	s = systemInfo;
 	do {
-		int cvar_flags;
+		unsigned int cvar_flags;
 
 		s = Info_NextPair( s, key, value );
 		if ( key[0] == '\0' ) {
@@ -432,7 +432,7 @@ void CL_SystemInfoChanged( qboolean onlyGame ) {
 			continue; // already processed
 		}
 
-		if ( ( cvar_flags = Cvar_Flags( key ) ) == CVAR_NONEXISTENT )
+		if ( ( cvar_flags = Cvar_Flags( key ) ) == (unsigned int)CVAR_NONEXISTENT )
 			Cvar_Get( key, value, CVAR_SERVER_CREATED | CVAR_ROM );
 		else
 		{
@@ -706,7 +706,7 @@ static void CL_ParseDownload( msg_t *msg ) {
 	}
 
 	size = MSG_ReadShort ( msg );
-	if (size < 0 || size > sizeof(data))
+	if (size < 0 || (size_t)size > sizeof(data))
 	{
 		Com_Error(ERR_DROP, "CL_ParseDownload: Invalid size %d for download chunk", size);
 		return;
